+++ src/features/public/Donations/screens/Projects.tsx
@@ -1,20 +1,25 @@
+import { TextField } from '@material-ui/core';
 import dynamic from 'next/dynamic';
 import React, { ReactElement } from 'react';
 import ProjectLoader from '../../../common/ContentLoaders/Projects/ProjectLoader';
-import SearchIcon from './../../../../assets/images/icons/SearchIcon';
 import CancelIcon from './../../../../assets/images/icons/CancelIcon';
-import { TextField } from '@material-ui/core';
+import SearchIcon from './../../../../assets/images/icons/SearchIcon';
 import styles from './../styles/Projects.module.scss';
+
+const MapLayout = dynamic(() => import('./MapboxMap'), {
+  ssr: false,
+  loading: () => <p>Loading...</p>,
+});
 
 const AllProjects = dynamic(() => import('../components/AllProjects'), {
   ssr: false,
   loading: () => <ProjectLoader />,
 });
 interface Props {
-  projects: Array<any>;
+  projects: any;
 }
 
-function Projects({ projects }: Props): ReactElement {
+function ProjectsList({ projects }: Props): ReactElement {
   const [selectedTab, setSelectedTab] = React.useState('featured');
   const [searchMode, setSearchMode] = React.useState(false);
   const [searchValue, setSearchValue] = React.useState('');
@@ -64,94 +69,104 @@ function Projects({ projects }: Props): ReactElement {
   const FeaturedProjectsProps = {
     projects: featuredProjects,
   };
+
+  const ProjectsProps = {
+    projects: projects,
+  };
   return (
-    <div className={styles.container}>
-      <div className={styles.cardContainer}>
-        {searchMode ? (
-          <div className={styles.headerSearchMode}>
-            <div className={styles.searchIcon}>
-              <SearchIcon />
-            </div>
-
-            <div className={styles.searchInput}>
-              <TextField
-                fullWidth={true}
-                autoFocus={true}
-                placeholder="Search Projects"
-                onChange={(e) => setSearchValue(e.target.value)}
-              />
-            </div>
-            <div
-              className={styles.cancelIcon}
-              onClick={() => {
-                setSearchMode(false);
-                setSearchValue('');
-              }}
-            >
-              <CancelIcon />
-            </div>
-          </div>
-        ) : (
-          <div className={styles.header}>
-            <div className={styles.tabButtonContainer}>
-              <div
-                className={styles.tabButton}
-                onClick={() => setSelectedTab('featured')}
-              >
-                <div
-                  className={
-                    selectedTab === 'featured'
-                      ? styles.tabButtonSelected
-                      : styles.tabButtonText
-                  }
-                >
-                  Transparent Projects
-                </div>
-                {selectedTab === 'featured' ? (
-                  <div className={styles.tabButtonSelectedIndicator} />
-                ) : null}
+    <>
+      <MapLayout
+        {...ProjectsProps}
+        mapboxToken={process.env.MAPBOXGL_ACCESS_TOKEN}
+      />
+      <div className={styles.container}>
+        <div className={styles.cardContainer}>
+          {searchMode ? (
+            <div className={styles.headerSearchMode}>
+              <div className={styles.searchIcon}>
+                <SearchIcon />
               </div>
 
+              <div className={styles.searchInput}>
+                <TextField
+                  fullWidth={true}
+                  autoFocus={true}
+                  placeholder="Search Projects"
+                  onChange={(e) => setSearchValue(e.target.value)}
+                />
+              </div>
               <div
-                className={styles.tabButton}
-                onClick={() => setSelectedTab('all')}
+                className={styles.cancelIcon}
+                onClick={() => {
+                  setSearchMode(false);
+                  setSearchValue('');
+                }}
               >
-                <div
-                  className={
-                    selectedTab === 'all'
-                      ? styles.tabButtonSelected
-                      : styles.tabButtonText
-                  }
-                >
-                  All {projects.length} Projects
-                </div>
-                {selectedTab === 'all' ? (
-                  <div className={styles.tabButtonSelectedIndicator} />
-                ) : null}
+                <CancelIcon />
               </div>
             </div>
-
-            <div
-              className={styles.searchIcon}
-              onClick={() => setSearchMode(true)}
-            >
-              <SearchIcon />
-            </div>
-          </div>
-        )}
-        {/* till here is header */}
-        <div className={styles.projectsContainer}>
-          {searchValue !== '' ? (
-            <AllProjects {...SearchResultProjectsProps} />
-          ) : selectedTab === 'all' ? (
-            <AllProjects {...AllProjectsProps} />
           ) : (
-            <AllProjects {...FeaturedProjectsProps} />
+            <div className={styles.header}>
+              <div className={styles.tabButtonContainer}>
+                <div
+                  className={styles.tabButton}
+                  onClick={() => setSelectedTab('featured')}
+                >
+                  <div
+                    className={
+                      selectedTab === 'featured'
+                        ? styles.tabButtonSelected
+                        : styles.tabButtonText
+                    }
+                  >
+                    Transparent Projects
+                  </div>
+                  {selectedTab === 'featured' ? (
+                    <div className={styles.tabButtonSelectedIndicator} />
+                  ) : null}
+                </div>
+
+                <div
+                  className={styles.tabButton}
+                  onClick={() => setSelectedTab('all')}
+                >
+                  <div
+                    className={
+                      selectedTab === 'all'
+                        ? styles.tabButtonSelected
+                        : styles.tabButtonText
+                    }
+                  >
+                    All {projects.length} Projects
+                  </div>
+                  {selectedTab === 'all' ? (
+                    <div className={styles.tabButtonSelectedIndicator} />
+                  ) : null}
+                </div>
+              </div>
+
+              <div
+                className={styles.searchIcon}
+                onClick={() => setSearchMode(true)}
+              >
+                <SearchIcon />
+              </div>
+            </div>
           )}
+          {/* till here is header */}
+          <div className={styles.projectsContainer}>
+            {searchValue !== '' ? (
+              <AllProjects {...SearchResultProjectsProps} />
+            ) : selectedTab === 'all' ? (
+              <AllProjects {...AllProjectsProps} />
+            ) : (
+              <AllProjects {...FeaturedProjectsProps} />
+            )}
+          </div>
         </div>
       </div>
-    </div>
+    </>
   );
 }
 
-export default Projects;
+export default ProjectsList;
